requirements.txt
prefect>=2.14.0,<3.0.0
lakefs_client>=0.91.0
selenium
webdriver-manager
pandas
python-dotenv